refactor: User 계층 분리
