+++ feature/sample/src/main/java/com/teamyg/feature/sample/UserIntent.kt
@@ -1,0 +1,8 @@
+package com.teamyg.feature.sample
+
+import com.tjyg.core.ui.UiIntent
+
+sealed interface UserIntent : UiIntent {
+    object LoadUser : UserIntent
+    data class UpdateName(val name: String) : UserIntent
+}
