+++ feature/sample/src/main/java/com/teamyg/feature/sample/UserSideEffect.kt
@@ -1,0 +1,7 @@
+package com.teamyg.feature.sample
+
+import com.tjyg.core.ui.UiSideEffect
+
+sealed interface UserSideEffect : UiSideEffect {
+    data class ShowToast(val message: String) : UserSideEffect
+}
