+++ feature/sample/src/main/java/com/teamyg/feature/sample/UserState.kt
@@ -1,0 +1,7 @@
+package com.teamyg.feature.sample
+
+import com.tjyg.core.ui.UiState
+
+data class UserState(
+    val isLoading: Boolean = false, val userName: String = "", val error: String? = null
+) : UiState
