+++ feature/sample/src/main/java/com/teamyg/feature/sample/UserViewModel.kt
@@ -2,9 +2,6 @@ package com.teamyg.feature.sample
 
 import androidx.lifecycle.viewModelScope
 import com.tjyg.core.ui.BaseViewModel
-import com.tjyg.core.ui.UserIntent
-import com.tjyg.core.ui.UserSideEffect
-import com.tjyg.core.ui.UserState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -24,7 +21,7 @@ class UserViewModel @Inject constructor() : BaseViewModel<UserState, UserIntent,
         viewModelScope.launch {
             updateState { copy(isLoading = true) }
             updateState { copy(isLoading = false, userName = "TEAM YG") }
-            postSideEffect(UserSideEffect.showToast("불러오기 성공"))
+            postSideEffect(UserSideEffect.ShowToast("불러오기 성공"))
         }
     }
 }
